Partially fixed issue #6 : Unit test failure on Windows.
Modified the help of `showargs` program to output the arguments list to a file if the environment variable `SHOWARGS_OUTPUT` is defined to a filename/filepath.
Reimplemented `TestUtils.cpp::getArgumentsFromSystem()` by setting environment variable `SHOWARGS_OUTPUT` and grab the output instead of relying to creating a temporary script and running it and piping the output to a file. This strategy can fail when you try to escape non-shell characters like `%` in a batch file. In a batch file, the `%` character must be escaped with `%%`.